Fixed CLI README format, moved format_README function to library.py
