+++ generate.py
@@ -22,7 +22,6 @@ import fileinput
 import markdownify
 
 MARKDOWN_TITLES = {
-    "cli" : "Command Line Interface",
     "python" : "Python Library",
     'data-types' : 'Data Types',
     'public-api' : 'Import & Export API',
@@ -76,32 +75,32 @@ def rename_to_readme(directory):
                     os.path.join(f"{root}", file_name),
                     os.path.join(f"{root}", raw_file_name, "README.md"),
                 )
-                format_README(os.path.join(f"{root}", raw_file_name, "README.md"), MARKDOWN_TITLES)
+                library.format_README(os.path.join(f"{root}", raw_file_name, "README.md"), MARKDOWN_TITLES)
 
 
-def format_README(readme_file_path: str, markdown_titles: dict):
-    """
-    Reads in README files and changes the title based on W&B Eng defined 
-    markdown titles passed by a global dictionary.
+# def format_README(readme_file_path: str, markdown_titles: dict):
+#     """
+#     Reads in README files and changes the title based on W&B Eng defined 
+#     markdown titles passed by a global dictionary.
 
-    Args:
-        readme_file_path (str): filepath to README
-        markdown_titles (dict): the default titles generated by docugen/generate.py are 
-        listed as keys. Values are our preferred README filename.
-    """
+#     Args:
+#         readme_file_path (str): filepath to README
+#         markdown_titles (dict): the default titles generated by docugen/generate.py are 
+#         listed as keys. Values are our preferred README filename.
+#     """
     
-    with open(readme_file_path, 'r') as f:
-        first_line = f.readline().strip('# ').strip('\n')
+#     with open(readme_file_path, 'r') as f:
+#         first_line = f.readline().strip('# ').strip('\n')
 
-        if first_line in markdown_titles:
-            new_title = '# ' + markdown_titles[first_line]
+#         if first_line in markdown_titles:
+#             new_title = '# ' + markdown_titles[first_line]
 
-    with fileinput.FileInput(readme_file_path, inplace = True) as f:    
-        for line in f:
-            if f.isfirstline():
-                print(new_title, end='\n')
-            else:
-                print(line, end='')
+#             with fileinput.FileInput(readme_file_path, inplace = True) as f:    
+#                 for line in f:
+#                     if f.isfirstline():
+#                         print(new_title, end='\n')
+#                     else:
+#                         print(line, end='')
 
 
 def clean_names(directory):
