+++ library.py
@@ -2,6 +2,7 @@ import configparser
 from importlib import import_module
 import operator
 import os
+import fileinput
 
 import wandb
 
@@ -26,6 +27,29 @@ subconfigs = util.process_subconfigs(config, subconfig_names)
 
 WANDB_CORE, WANDB_DATATYPES, WANDB_API, WANDB_INTEGRATIONS = subconfigs
 
+def format_README(readme_file_path: str, markdown_titles: dict):
+    """
+    Reads in README files and changes the title based on W&B Eng defined 
+    markdown titles passed by a global dictionary.
+
+    Args:
+        readme_file_path (str): filepath to README
+        markdown_titles (dict): the default titles generated by docugen/generate.py are 
+        listed as keys. Values are our preferred README filename.
+    """
+    
+    with open(readme_file_path, 'r') as f:
+        first_line = f.readline().strip('# ').strip('\n')
+
+        if first_line in markdown_titles:
+            new_title = '# ' + markdown_titles[first_line]
+
+            with fileinput.FileInput(readme_file_path, inplace = True) as f:    
+                for line in f:
+                    if f.isfirstline():
+                        print(new_title, end='\n')
+                    else:
+                        print(line, end='')
 
 def build(commit_id, code_url_prefix, output_dir):
     """Builds docs in stages: main library, then subcomponents."""
